feat: add full webpack config, refactor folder structure
