+++ .gitignore
@@ -1,1 +1,2 @@
+dist
 node_modules
