+++ package.json
@@ -1,42 +1,59 @@
 {
-  "name": "codejam-image-api",
+  "name": "webpack-boilerplate",
   "version": "1.0.0",
-  "description": "",
-  "main": "script.js",
-  "dependencies": {
-    "canvas": "^2.6.0",
-    "chai": "^4.2.0",
-    "eslint": "^6.6.0",
-    "jsdom": "^15.2.1",
-    "jsdom-global": "^3.0.2"
+  "description": "WebPack: HTML, CSS, SCSS, JS, Img, Fonts",
+  "main": ".eslintrc.js",
+  "browserslist": [
+    "> 0.25%, not dead"
+  ],
+  "scripts": {
+    "dev": "webpack-dev-server --history-api-fallback --inline --progress --open",
+    "build": "npm run clear -s && webpack --progress",
+    "lint": "eslint ./**/*.js",
+    "clear": "del-cli dist"
+  },
+  "author": "Netizen",
+  "license": "ISC",
+  "repository": {
+    "type": "git",
+    "url": "git+https://github.com/Net-zen/webpack-boilerplate.git"
   },
   "devDependencies": {
+    "@babel/core": "^7.6.0",
+    "@babel/polyfill": "^7.7.0",
+    "@babel/preset-env": "^7.6.0",
+    "autoprefixer": "^9.6.1",
+    "babel-loader": "^8.0.6",
+    "copy-webpack-plugin": "^5.0.5",
     "css-loader": "^3.2.0",
+    "css-mqpacker": "^7.0.0",
+    "cssnano": "^4.1.10",
+    "del-cli": "^3.0.0",
+    "eslint": "^6.6.0",
+    "eslint-config-airbnb-base": "^14.0.0",
+    "eslint-config-prettier": "^6.5.0",
+    "eslint-loader": "^3.0.2",
+    "eslint-plugin-import": "^2.18.2",
     "file-loader": "^4.2.0",
     "html-loader": "^0.5.5",
     "html-webpack-plugin": "^3.2.0",
+    "image-webpack-loader": "^6.0.0",
     "mini-css-extract-plugin": "^0.8.0",
-    "mocha": "^6.2.2",
     "node-sass": "^4.13.0",
+    "postcss-loader": "^3.0.0",
     "sass-loader": "^8.0.0",
     "style-loader": "^1.0.0",
+    "uglifyjs-webpack-plugin": "^2.2.0",
     "webpack": "^4.41.2",
-    "webpack-cli": "^3.3.10",
+    "webpack-cli": "^3.3.9",
     "webpack-dev-server": "^3.9.0"
   },
-  "scripts": {
-    "build": "webpack",
-    "start": "webpack-dev-server"
-  },
-  "repository": {
-    "type": "git",
-    "url": "git+https://github.com/ReaziS/codejam-image-api.git"
-  },
-  "author": "",
-  "license": "ISC",
   "bugs": {
-    "url": "https://github.com/ReaziS/codejam-image-api/issues"
+    "url": "https://github.com/Net-zen/webpack-boilerplate/issues"
   },
-  "homepage": "https://github.com/ReaziS/codejam-image-api#readme",
-  "keywords": []
+  "homepage": "https://github.com/Net-zen/webpack-boilerplate#readme",
+  "keywords": [],
+  "dependencies": {
+    "core-js": "^2.6.10"
+  }
 }
