+++ postcss.config.js
@@ -1,0 +1,21 @@
+// autoprefixer - https://github.com/postcss/autoprefixer
+// css-mqpacker - https://github.com/hail2u/node-css-mqpacker
+// cssnano      - https://github.com/hail2u/node-css-mqpacker
+
+// npm install postcss-loader autoprefixer css-mqpacker cssnano --save-dev
+
+module.exports = {
+  plugins: [
+    require('autoprefixer'),
+    require('css-mqpacker'),
+    require('cssnano')({
+      preset: [
+        'default', {
+          discardComments: {
+            removeAll: true
+          }
+        }
+      ]
+    })
+  ]
+}
