+++ src/css/style.scss
@@ -1,6 +1,5 @@
 
 body {
-    display: none;
     margin: 0 auto;
     max-width: 1440px;
 }
