+++ src/index.html
@@ -4,42 +4,41 @@
     <meta charset="UTF-8">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
     <meta http-equiv="X-UA-Compatible" content="ie=edge">
-    <link rel="stylesheet" href="../src/assets/styles/style.css">
     <title>Document</title>
 </head>
 <body>
     <header>
         <div class="burger">
-            <img src="assets/images/Shape.svg" alt="">
+            <img src="images/Shape.svg" alt="">
             <span class="header__title">CodeJam - Pallete</span>
         </div>
-        <div class="options"><img src="../src/assets/images/menu icon.svg" alt=""></div>
+        <div class="options"><img src="images/menu icon.svg" alt=""></div>
     </header>
     <section class="draw-wrapper">
         <div class="instruments-wrapper">
                 <div class="instruments-main" id="instruments-panel">
                     <div class="instrument paint-bucket" id="paintBucket">
-                        <img src="assets/images/Vector.svg" alt="">
+                        <img src="images/Vector.svg" alt="">
                         <span>Paint bucket</span>
                     </div>
                     <div class="instrument color-picker" id="colorPicker">
-                        <img src="assets/images/Vector2.svg" alt="">
+                        <img src="images/Vector2.svg" alt="">
                         <span>Choose color</span>
                     </div>
                     <div class="instrument pencil" id="pencil">
-                        <img src="assets/images/icon-copy.svg" alt="">
+                        <img src="images/icon-copy.svg" alt="">
                         <span>Pencil</span>
                     </div>
                     <div class="instrument" id="bresenham">
-                        <img src="assets/images/icon-copy.svg" alt="">
+                        <img src="images/icon-copy.svg" alt="">
                         <span>Bresenham</span>
                     </div>
                     <div class="instrument move">
-                        <img src="assets/images/Move.svg" alt="">
+                        <img src="images/Move.svg" alt="">
                         <span>Move</span>
                     </div>
                     <div class="instrument transform">
-                        <img src="assets/images/Vector1.svg" alt="">
+                        <img src="images/Vector1.svg" alt="">
                         <span>Transform</span>
                     </div>
                 </div>
@@ -75,9 +74,7 @@
                     <input type="range" id="imageSize" min="1" max="8" step="1">
                     <div class="grayscale__button" id="grayscale">Grayscale</div>
                 </div>
-               
             </div>
     </section>
-    <script type="text/javascript" src="index.js"></script>
 </body>
 </html>
