+++ webpack.config.js
@@ -1,44 +1,152 @@
-const HtmlWepPackPlugin = require('html-webpack-plugin');
-const MiniCssExtractPlugin = require('mini-css-extract-plugin');
+const path = require('path');
+const HtmlWebPackPlugin = require('html-webpack-plugin');
+const MiniCssExtractPlugin = require("mini-css-extract-plugin");
+const UglifyJSPlugin = require('uglifyjs-webpack-plugin');
 
-module.exports = {
+const webpack = require('webpack');
+
+const ENV = process.env.npm_lifecycle_event;
+const isDev = ENV === 'dev';
+const isProd = ENV === 'build';
+
+function setDevTool() {
+  if (isDev) {
+    return 'cheap-module-eval-source-map';
+  } else {
+    return 'none';
+  }
+}
+
+function setDMode() {
+  if (isProd) {
+    return 'production';
+  } else {
+    return 'development';
+  }
+}
+
+const config = {
+  target: "web", 
+  entry: './src/js/index.js',
+  output: {
+    path: path.resolve(__dirname, 'dist'),
+    filename: '[name].js'
+  },
+  mode: setDMode(),
+  devtool: setDevTool(),
   module: {
-    rules: [
-      {
+    rules: [{
         test: /\.html$/,
+        use: [{
+          loader: 'html-loader',
+          options: {
+            minimize: false
+          }
+        }]
+      },
+      {
+        test: /\.js$/,
+        use: ['babel-loader'/* , 'eslint-loader' */],
+        exclude: [
+          /node_modules/
+        ]
+      },
+      {
+        test: /\.css$/,
         use: [
+          'style-loader',
+          MiniCssExtractPlugin.loader,
           {
-            loader: 'html-loader',
+            loader: 'css-loader',
             options: {
-              minimize: true,
-            },
-          },
-        ],
+              sourceMap: true
+            }
+          }, {
+            loader: 'postcss-loader',
+            options: { sourceMap: true, config: { path: './postcss.config.js' } }
+          }
+        ]
       },
       {
         test: /\.scss$/,
         use: [
           'style-loader',
-          'css-loader',
-          'sass-loader',
-        ],
+          MiniCssExtractPlugin.loader,
+          {
+            loader: 'css-loader',
+            options: {
+              sourceMap: true
+            }
+          }, {
+            loader: 'postcss-loader',
+            options: { sourceMap: true, config: { path: './postcss.config.js' } }
+          }, {
+            loader: 'sass-loader',
+            options: {
+              sourceMap: true
+            }
+          }
+        ]
       },
       {
-        test: /\.(png|svg|jpg|gif)$/,
+        test: /\.(jpg|png|svg|gif)$/,
         use: [
-          'file-loader',
-        ],
+          {
+            loader: 'file-loader',
+            options: {
+              outputPath: 'img',
+              name: '[name].[ext]'
+            }},
+          {
+            loader: 'image-webpack-loader',
+            options: {
+              mozjpeg: {
+                processive: true,
+                quality: 98
+              }
+            }
+          }
+        ]
       },
-    ],
+      {
+        test: /\.(woff|woff2|ttf|otf|eot)$/,
+        use: [{
+          loader: 'file-loader',
+          options: {
+            outputPath: 'fonts'
+          }
+        }]
+      }
+    ]
   },
+
   plugins: [
-    new HtmlWepPackPlugin({
-      template: './src/canvas.html',
-      filename: './index.html',
-    }),
     new MiniCssExtractPlugin({
-      filename: '[name].css',
-      chunkFilename: '[id].css',
+      filename: 'style.css',
     }),
+    new HtmlWebPackPlugin({
+      template: './src/index.html',
+      filename: './index.html'
+    }),
+      new webpack.DefinePlugin({
+      API_KEY: JSON.stringify(process.env.API_KEY),
+      APP_ENV: JSON.stringify(process.env.APP_ENV)
+    })
   ],
+
+  devServer: {
+    contentBase: path.join(__dirname, 'dist'),
+    compress: true,
+    port: 3000,
+    stats: 'errors-only',
+    clientLogLevel: 'none'
+  }
+}
+
+if (isProd) {
+  config.plugins.push(
+    new UglifyJSPlugin(),
+  );
 };
+
+module.exports = config;
